requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langserve
langchain-huggingface
python-dotenv
openai